#1 가격 정보 타입 double로 변경

+++ module-analyzer/src/main/java/analyzer/domain/stock/dto/StockData.java
@@ -7,7 +7,7 @@ public record StockData(
     String ticker,
 
     @JsonProperty("trade_price")
-    int price
+    double price
 ) {
 
 }
